Refactor data source models into models package
