+++ datasources/models/__init__.py
@@ -1,0 +1,16 @@
+"""
+This package contains the :class:`DataSource` model representing an internal or
+external data source and the models necessary to support it.
+"""
+
+from .datasource import (
+    Licence,
+    UserPermissionLevels,
+    UserPermissionLink,
+    DataSource
+)
+
+from .metadata import (
+    MetadataField,
+    MetadataItem
+)
